requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
openai==1.54.4
httpx==0.28.0
python-dotenv==1.0.1
pydantic==2.10.2
python-multipart==0.0.12
greenlet==3.1.1
weasyprint==62.3
jinja2==3.1.4
markdown==3.7
aiohttp==3.10.11
aiofiles==24.1.0
python-pptx==1.0.2